pyproject.toml
[build-system]
requires = ["uv_build >= 0.7.19, <0.9.0"]
build-backend = "uv_build"

[project]
name = "pageextractor"
version = "0.1.6"
description = "Use SAM2 to extract and warp a page from a photo."
readme = "README.md"
authors = [
  { name = "Herbert Kruitbosch", email = "H.T.Kruitbosch@rug.nl" },
]
requires-python = ">=3.10"
license = "Apache-2.0"
license-files = ["LICENSE"]
dependencies = [
    "gradio>=5.29.0",
    "litserve>=0.2.8",
    "opencv-python-headless>=4.10.0.84",
    "pillow>=10.2.0",
    "rasterio>=1.4.3",
    "sam2",
    "scikit-image>=0.24.0",
    "shapely>=2.1.1",
    "supervision>=0.23.0",
    "torch>=2.3.1",
    "torchvision>=0.18.1",
    "tqdm>=4.67.1",
    "transformers>=4.44.2",
]


[project.urls]
Homepage = "https://github.com/UG-Team-Data-Science/pageextractor"
Issues = "https://github.com/UG-Team-Data-Science/pageextractor/issues"

[tool.ruff]
target-version = "py311"
line-length = 120
fix = true
select = [
    # https://github.com/charliermarsh/ruff#pyflakes-f
    "F", # Pyflakes
    # https://github.com/charliermarsh/ruff#pycodestyle-e-w
    "E", # pycodestyle
    "W", # Warning
    # https://github.com/charliermarsh/ruff#flake8-comprehensions-c4
    # https://github.com/charliermarsh/ruff#mccabe-c90
    "C", # Complexity (mccabe+) & comprehensions
    # https://github.com/charliermarsh/ruff#pyupgrade-up
    "UP", # pyupgrade
    # https://github.com/charliermarsh/ruff#isort-i
    "I", # isort
]
ignore = [
    # https://github.com/charliermarsh/ruff#pycodestyle-e-w
    "E402", # module level import not at top of file
    # https://github.com/charliermarsh/ruff#pyupgrade-up
    "UP006", # use-pep585-annotation
    "UP007", # use-pep604-annotation
    "E741", # Ambiguous variable name
]
[tool.ruff.per-file-ignores]
"__init__.py" = [
    "F401",  # unused import
    "F403",  # star imports
    ]

[tool.ruff.mccabe]
max-complexity = 24

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.uv.sources]
sam-2 = { git = "https://github.com/facebookresearch/segment-anything-2", rev = "c2ec8e14a185632b0a5d8b161928ceb50197eddc" }
